requirements.txt
numpy
joblib
sklearn
newspaper3k